+++ tests/test-config-local.jasmine.ts
@@ -1,9 +1,9 @@
-import "jasmine";
+// import "jasmine";
 import fs from "fs";
 import path from "path";
+import { rimraf } from "rimraf";
 import { Storage } from "../src/Storage";
 import { StorageType, ConfigLocal } from "../src/types";
-import { promiseRimraf } from "./util";
 
 // describe("test jasmine", () => {
 //   it("weird", () => {
@@ -13,8 +13,19 @@ import { promiseRimraf } from "./util";
 // });
 
 describe(`testing local urls`, () => {
+  beforeAll(async () => {
+    // override Linux's umask 0o002
+    process.umask(0o000);
+  });
+
   it("[0]", async () => {
-    const storage = new Storage("local://tests/tmp/the-buck");
+    const storage = new Storage("local://tests/tmp/the-buck?param=value");
+    await storage.init();
+    expect(storage.getType()).toBe(StorageType.LOCAL);
+    expect(storage.getSelectedBucket()).toBe("the-buck");
+  });
+  it("[0a]", async () => {
+    const storage = new Storage("local://the-buck?param=value");
     await storage.init();
     expect(storage.getType()).toBe(StorageType.LOCAL);
     expect(storage.getSelectedBucket()).toBe("the-buck");
@@ -39,17 +50,18 @@ describe(`testing local urls`, () => {
     });
     await storage.init();
     expect(storage.getType()).toBe(StorageType.LOCAL);
-    expect(storage.getSelectedBucket()).toBe("the-buck");
+    expect(storage.getSelectedBucket()).toBe("");
   });
 
   it("[4]", async () => {
     const storage = new Storage({
       type: StorageType.LOCAL,
       directory: "tests/tmp",
+      bucketName: "the-buck",
     });
     await storage.init();
     expect(storage.getType()).toBe(StorageType.LOCAL);
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.getSelectedBucket()).toBe("the-buck");
   });
 
   it("[5] numeric values in options stay numeric and keep their radix (8)", async () => {
@@ -78,8 +90,8 @@ describe(`testing local urls`, () => {
     expect(storage.getSelectedBucket()).toBe("tmp");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe("0o777");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
-    expect(mode).toBe(16877);
-    await promiseRimraf(path.join(process.cwd(), "tests", "tmp"));
+    expect(mode.toString(8)).toBe("40777");
+    await rimraf(path.join(process.cwd(), "tests", "tmp"));
   });
 
   it("[6a]", async () => {
@@ -88,63 +100,65 @@ describe(`testing local urls`, () => {
     expect(storage.getSelectedBucket()).toBe("tmp");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe("777");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
-    expect(mode).toBe(16877);
-    await promiseRimraf(path.join(process.cwd(), "tests", "tmp"));
+    expect(mode.toString(8)).toBe("40777");
+    await rimraf(path.join(process.cwd(), "tests", "tmp"));
   });
 
   it("[6c] octal numbers get converted to radix 10", async () => {
     const storage = new Storage({
       type: StorageType.LOCAL,
       directory: "tests/tmp",
+      bucketName: "the-buck",
       mode: 0o777,
     });
     await storage.init();
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.getSelectedBucket()).toBe("the-buck");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe(511);
-    const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
-    expect(mode).toBe(16877);
-    await promiseRimraf(path.join(process.cwd(), "tests", "tmp"));
+    const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp", "the-buck")))
+      .mode;
+    expect(mode.toString(8)).toBe("40777");
+    await rimraf(path.join(process.cwd(), "tests", "tmp"));
   });
 
   it("[6d]", async () => {
     const storage = new Storage({
       type: StorageType.LOCAL,
       directory: "tests/tmp",
+      bucketName: "the-buck",
       mode: 511,
     });
     await storage.init();
-    expect(storage.getSelectedBucket()).toBe("tmp");
+    expect(storage.getSelectedBucket()).toBe("the-buck");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe(511);
-    const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
-    expect(mode).toBe(16877);
-    await promiseRimraf(path.join(process.cwd(), "tests", "tmp"));
+    const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp", "the-buck")))
+      .mode;
+    expect(mode.toString(8)).toBe("40777");
+    await rimraf(path.join(process.cwd(), "tests", "tmp"));
   });
 
   it("[6e]", async () => {
-    const storage = new Storage({
-      type: StorageType.LOCAL,
-      directory: "tests/tmp",
-      mode: "0o777",
-    });
+    const storage = new Storage("local://tests/tmp?mode=0o777");
     await storage.init();
     expect(storage.getSelectedBucket()).toBe("tmp");
     expect((storage.getConfiguration() as ConfigLocal).mode).toBe("0o777");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
-    expect(mode).toBe(16877);
-    await promiseRimraf(path.join(process.cwd(), "tests", "tmp"));
+    expect(mode.toString(8)).toBe("40777");
+    await rimraf(path.join(process.cwd(), "tests", "tmp"));
   });
 
+  // pass octal value as string
   it("[6f]", async () => {
     const storage = new Storage({
       type: StorageType.LOCAL,
-      directory: "tests/tmp",
-      mode: "777",
+      directory: "tests",
+      bucketName: "tmp",
+      mode: "755", // this is an error! the parseMode function will return the default value 0o777
     });
     await storage.init();
     expect(storage.getSelectedBucket()).toBe("tmp");
-    expect((storage.getConfiguration() as ConfigLocal).mode).toBe("777");
+    expect((storage.getConfiguration() as ConfigLocal).mode).toBe("755");
     const mode = (await fs.promises.stat(path.join(process.cwd(), "tests", "tmp"))).mode;
-    expect(mode).toBe(16877);
-    await promiseRimraf(path.join(process.cwd(), "tests", "tmp"));
+    expect(mode.toString(8)).toBe("40777");
+    await rimraf(path.join(process.cwd(), "tests", "tmp"));
   });
 });