+++ tests/util.ts
@@ -33,3 +33,26 @@ export const copyFile = (
         resolve();
       });
   });
+
+export const parseMode = (s: number | string): string | number => {
+  // if mode is a number, parseMode assumes it is a decimal number
+  if (typeof s === "number") {
+    if (s < 0) {
+      throw new Error(
+        `The argument 'mode' must be a 32-bit unsigned integer or an octal string. Received ${s}`
+      );
+    }
+    return s;
+  }
+
+  // mode is a string
+
+  // e.g "0x755" (octal)
+  if (s.startsWith("0o")) {
+    return parseInt(s.substring(2), 8).toString(8);
+  }
+  // e.g '511' (decimal)
+  const i = parseInt(s, 10);
+  // quick fix for erroneously passed octal number as string (without 0o prefix)
+  return i > 511 ? 511 : i;
+};